:bug: Fix use of 1Password SSH Agent in WSL

+++ conf.d/1password.fish
@@ -5,8 +5,10 @@ end
 
 # 1Password SSH Agent
 if not set -q SSH_AUTH_SOCK
-  ln -sf ~/.1password/agent.sock ~/.agent.sock
-  set -x SSH_AUTH_SOCK ~/.agent.sock
+  if test -S ~/.1password/agent.sock
+    ln -sf ~/.1password/agent.sock ~/.agent.sock
+    set -x SSH_AUTH_SOCK ~/.agent.sock
+  end
 else
   ln -sf "$SSH_AUTH_SOCK" ~/.agent.sock
 end
